Remove dependencies of dependencies
It's usually a good idea to only keep track of the packages/versions that you intend to use. That way if there's updates to the dependencies, they aren't locked down when you didn't even install that version yourself anyways. It's also more clear for projects like this which packages are being used directly.